Implementação da alteração de restaurante

+++ src/main/java/br/com/guimassolucoes/guimasfood/infrastructure/repository/EstadoRespositoryImpl.java
@@ -5,7 +5,9 @@ import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
 
 import br.com.guimassolucoes.guimasfood.domain.model.Estado;
 import br.com.guimassolucoes.guimasfood.domain.repository.EstadoRepository;
@@ -19,6 +21,7 @@ public class EstadoRespositoryImpl implements EstadoRepository {
 	@Override
 	public List<Estado> todos() {
 		List<Estado> result = em.createQuery("from Estado", Estado.class).getResultList();
+		
 		return result;
 	}
 
@@ -27,14 +30,21 @@ public class EstadoRespositoryImpl implements EstadoRepository {
 		return em.find(Estado.class, id);
 	}
 
+	@Transactional
 	@Override
 	public Estado salvar(Estado estado) {
 		return em.merge(estado);
 	}
 
+	@Transactional
 	@Override
 	public void remover(Estado estado) {
 		estado = porId(estado.getId());
+
+		if(estado==null) {
+			throw new EmptyResultDataAccessException(1);
+		}
+		
 		em.remove(estado);
 	}
